requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-decouple==3.8
pillow==10.1.0
aiofiles==23.2.1
stripe==7.7.0
email-validator==2.1.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
httpx==0.25.2
pandas==2.2.0
openpyxl==3.1.2

# Additional dependencies for notifications
python-dotenv==1.0.0
requests==2.31.0 